Refactor import paths for bookmark management components
- Updated import statements in create-bookmark.tsx, manage-groups.tsx, search-bookmarks.tsx, authenticated-view.tsx, and use-bookmarks.ts to use relative paths for the lib directory.
- This change improves the organization and clarity of the codebase by ensuring consistent import paths across the project.
