+++ extensions/bmrks/src/components/authenticated-view.tsx
@@ -1,5 +1,5 @@
 import { Action, ActionPanel, Detail, Icon, List, openExtensionPreferences } from "@raycast/api";
-import { useAuth } from "../../lib/use-auth";
+import { useAuth } from "../lib/use-auth";
 import { User } from "@supabase/supabase-js";
 
 export default function AuthenticatedView({
